Allow D2 to initialize before setting up the Redux Store
When hydrating the persisted state, e.g. in the situation where:

    1. User inputs events offline
    2. User closes tab with capture app
    3. User re-opens capture app

We have the state persisted, and as soon as the Store is initialised
redux-offline will try to run the events in the outbox to sync them. At
this point D2 *has* to be available or the App will crash.

The `initialize` function does not depend on any state so it can be run
before the `getStore` functions. We still get the loader while d2 loads.