chore(release): stop support for 2.39 (#4131)
To keep a manageable codebase, we have decided to support only the same version range as the backend. However, in contrast to the backend, we will release new features for all the supported versions as long as we have the backend support.

BREAKING CHANGE: Bump version to 103.0.0 to facilitate potential fixes for 2.39.